+++ src/main/java/insane96mcp/enhancedai/ai/EAHurtByTargetGoal.java
@@ -1,5 +1,6 @@
 package insane96mcp.enhancedai.ai;
 
+import insane96mcp.enhancedai.modules.mobs.targeting.Targeting;
 import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.ai.goal.target.TargetGoal;
 import net.minecraft.world.entity.ai.targeting.TargetingConditions;
@@ -32,8 +33,8 @@ public class EAHurtByTargetGoal extends TargetGoal {
 		LivingEntity currentTarget = this.mob.getTarget();
 		if (currentTarget != null && currentTarget == hypotheticalNewMob)
 			return false;
-		//New check to prefer players oven non-player entities
-		if (currentTarget instanceof Player && !(hypotheticalNewMob instanceof Player))
+		//New check to prefer players oven non-player entities if enabled
+		if (currentTarget instanceof Player && !(hypotheticalNewMob instanceof Player) && Targeting.preferPlayers)
 			return false;
 
 		//New check to not switch target if the current one is closer