+++ changelog.md
@@ -4,9 +4,10 @@
 * Added a new explosion sound for Angry Creepers. It's now the default
     * The fuse sound is vanilla but the explosion sound is the old minecraft beta one
 * Configurable Animal flee speed
+* Added a config option for Better Hurt by target to prefer players to other entities
 * Performance improvement for fisher mobs
 
-## 2.4.10
+## 2.4.10 
 * Removed `use_follow_range_changes` entity type tag
   * Replaced with `change_follow_range` that applies only the follow range override
   * Also with `apply_xray` that applies only the xray range override